Small reformat

+++ lusuguo.cpp
@@ -62,15 +62,15 @@
 LuSuGuo::LuSuGuo(size_t vertexCount) : Model(vertexCount) {
     graph.addVertex();
 
-    for (unsigned long int id = 1; id < vertexCount; id++) {
+    for (size_t id = 1; id < vertexCount; id++) {
         int l = (int)std::floor(std::log2(id + 1));
         graph.addVertex();
 
         if (l > 1) {
-            graph.addEdge(id, (unsigned long int)std::floor((id - 1) / 2) % (l - 1));
+            graph.addEdge(id, (size_t)std::floor((id - 1) / 2) % (l - 1));
         }
 
-        graph.addEdge(id, (unsigned long int)std::floor((id + 1) / 2) - 1);
+        graph.addEdge(id, (size_t)std::floor((id + 1) / 2) - 1);
 
         if (id % 2 == 0) {
             graph.addEdge(id, id - 1);
